#45 update file structure, update tests
