+++ src/engine/timer.js
